pyproject.toml
[project]
name = "phone-number-info"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "django>=6.0.1",
    "django-apscheduler>=0.7.0",
    "django-environ>=0.12.0",
    "django-ninja-extra>=0.30.8",
    "psycopg>=3.3.2",
    "uvicorn[standard]>=0.40.0",
]
